SFX Stuff
- Added falling noise
- Added player-getting-hurt noise
- Added monsters-getting-hurt noises
- Added different SFX for hitting hard surfaces & trees
The noises are terrible. But they signal something: a skeleton upon
which we can give props and monsters a noise to make upon being hit. In
addition to merely adding these sounds, I have created functions that
enable fairly fine-grained control over what noises do or don't play
when X or Y gets hit by this or that weapon.